Introducing bloc
