+++ lib/bloc/base_bloc.dart
@@ -1,5 +1,10 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_ecoapp/models/base.dart';
 
-class BaseBloc<T extends BaseModel>{
+abstract class BaseBloc<T extends BaseModel> extends Bloc{
+  BaseBloc(initialState) : super(initialState);
+
+  @override
+  Stream mapEventToState(event);
 
 }
